Highscore implementation
The player's time alive will be kept as "BestTime" on the initial play, and the float will be saved as "BestTimeFloat" for checking if it is beaten. If the player survived for longer than the current Best Time, (checked by comparing the "BestTimeFloat" to the totalTime float on player death) it will be overwritten. The current best time is shown on screen in the Main Menu Scene.